Streamline SNMP variable class definition
Before, there was a netsnmpVariable class and Integer32, Unsigned32, DisplayString32 etc. classes deriving from it (and an intermediate class netsnmpIntegerVariable). The variables carried the SNMP registration intelligence. In addition, netsnmpAgent had an addVar() method and proxy methods named liked the classes, taking care of creating variable instance and adding them to an agent's internal variable registry.

Now, netsnmpAgent features a VarTypeClass decorator method that gets applied to methods Integer32, Unsigned32, DisplayString32 etc. which do nothing but return a dictionary with variable type properties. The decorator transforms these functions so that they a.) define the suitable class b.) create an instance and c.) register that instance with net-snmp. The classes themselves are now nothing but wrappers for the ctypes variables. So the registration intelligence (and oidstr2oid()) is back in netsnmpAgent itself again.-