Generalize for SNMP tables, separate SNMP object creation and registration
This change generalizes netsnmpagent in that we now deal with "SNMP objects" instead of just SNMP variables. This is in preparation of the upcoming support of SNMP tables.

Because of the semantics associated with how we will wrap tables, SNMP object creation and registration under a certain OID are now separate tasks. SNMP objects now must be explicitly registered (see example_agent.py).

Further details of the changes:
- Because registering a certain SNMP object is object-dependent, the VarTypeClass() decorator now expects a new member "register_func" that specifies the callback function that will be called by the new generic register() method.
- register() incorporates the former oidstr2oid() method
- The VarTypeClass() decorator's variable registration code went into register() and the new callback function _registerWatcher() as well.
- Variable type class instances now possess most of the atttributes from their props[] array such as "asntype" and "flags"
- _data_size and _max_size were not always set correctly
- Variable type class instances now possess a new cref() method which always returns the correct pointer to the enclosed _cvar, depending on the variable type