API change: swap "oidstr" and "initval" arguments for scalar variables
The function signature for create_vartype_class() has changed and as such the constructor
method signature for all classes using the @VarTypeClass decorator, that is, all classes
representing SNMP scalar variables, eg. Integer32(), Unsigned32() etc. "initval" now
comes first so that in scenarios where we just a SNMP variable without the intent of
registering it we can simply write

  agent.Integer32(20)

instead of

  agent.Integer32(initval=20)

which, of course, implies "oidstr" == None. The former construct is of much more use when
being interpreted as "initval" instead of "oidstr" as in

  agent.Integer32("EXAMPLE-MIB::exampleInteger32")

which continues requiring an "oidstr=" prefix to work.

This means that if you have already written agents using python-netsnmpagent and you did
not follow the style in example_agent.py with explicit "initval=" and "oidstr=" qualifiers,
you WILL have to adapt your code.