Replace hardcoded version number with latest Git tag
With this change, the version number is no longer hard-coded inside setup.py. Instead, the
version number is derived from the latest git tag using "git describe". The Makefile (which
is now more than a pure convenience Makefile) will generate setup.py from setup.py.in
(which has been renamed accordingly), so that a source distribution (.tar.gz) will have
the correct version number inside.

As a special rule, whenever the Git checkout is not at a tag, "git describe" will generate
a string similar to "0.2-21-g0e10fca", indicating that there have been 21 commits since the
last tag "0.2" and the latest commit is g0e10fca. We translate this to "0.2_next21_g0e10fca"
so a.) RPM does not complain about the dashes (it doesn't allow them in its "Version" field)
and b.) the "next" indicates that this is a forthcoming version.