Generate RPM spec file along with source distribution archive
Utility of the spec file is increased if it has python-netsnmpagent's version number
already specified -- the old way of requiring a RPM "--define" only worked when the
Git repo was available. Now one can take the generated dist/python-netsnmpagent.spec
along with the source archive and simply copy it some build host and build.