Automatically generate %changelog section for RPM spec file
The spec file uses a different changelog style than the ChangeLog file, which would be too
verbose for this purpose. Usually package maintainers note changes to the spec file itself
here as well and often give a more high-level summary of changes of the software being
packaged. Our automatically generated changelog will not be of the same quality as we only
indicate the timestamps a version was tagged and an "Update to version x.y.z"-style message.
But until now we didn't use the ChangeLog section at all. And package maintainers can still
replace it with a more detailed, manually maintained version.