Support detecting connection establishment/failure
So far we just relied on net-snmp resp. its init_snmp() function to establish
the AgentX connection to the master snmpd instance and neither knew nor cared
about whether a connection could be established at all. A mistake in the
"MasterSocket" option would only cause a warning on stderr, emitted by net-snmp
itself, only. Likewise, indication of a successful connection was through log
messages on stderr only, too.

This change introduces support for explicit detection of connection
establishment and failure. To this end, the netsnmpAgent class' internal
"_started" flag was replaced with a "_status" property that can take on the
values defined in the new "netsnmpAgentStatus" enumeration. Your code should
however NOT access this property -- just make sure to catch the new
netsnmpAgentExceptions (implemented in the example agents in cd4ee23b).

The change is quite large due to the fact that net-snmp itself in neither the
5.4.x nor 5.7.x versions actually returns status information: init_snmp() is
defined as "void", lending due to the fact that it will actually trigger a
sequence of function calls including callback functions that return "void"
themselves. Unfortunately there is also no dedicated callback mechanism for
client code such as python-netsnmpagent that gets called on connection
status changes. So we had to invent it on our own, using two callback
functions as more or less ugly workarounds: a custom log handler and, for
net-snmp 5.4.x support, another callback function that abuses a callback
hook originally intended for different purposes. See the sources for details,
they are well commented.

In this version, our custom log handler writes all log messages it receives
to stderr, resembling previous behavior with the difference that we also log
a message's priority level as defined by net-snmp, eg. "[Info] Foo.".