netsnmpapi: define netsnmp_init_mib() instead of unused init_mib()
In netsnmpapi we defined init_mib() but in netsnmpagent we called
netsnmp_init_mib() which was undefined in netsnmpapi. That didn't cause trouble
because a.) ctypes implicitly treated it as being "void netsnmp_init_mib(void)"
and b.) by coincidence that matched the function signature.

The confusion stemmed from a function rename in the net-snmp project, see
https://www.mail-archive.com/net-snmp-coders%40lists.sourceforge.net/msg08417.html