netsnmpagent: Call net-snmp's shutdown_agent() in shutdown() method
Properly written agents should call shutdown_agent() in addition to
snmp_shutdown() as we did a init_agent() before the init_snmp() call.

Note: with ALL net-snmp versions up to and including the current 5.7.3 beta
this will still NOT allow for proper cleanup in such a way that you could
"del" the netsnmpAgent instance, create a new one and have that one connect to
the master agent successfully. Even with 5.7.x's code cleanups and the subsequent
removal of "#ifdef SHUTDOWN_AGENT_CLEANLY" in snmpd.c, it still seems to be
necessary to exit the process (or thread) and have the OS do additional
cleanup work if a substantially different netsnmpAgent instance (eg. with a
different MasterSocket) is to be created. Unfortunately this also affects
test cases where one wants to run each test in a defined, clean and
reproducible environment.