netsnmptestenv: Fix snmpcmd() error detection/exceptions
For one thing, we used re.match() to search for matches at the end of the
output, thus exceptions were never raid. Second, we only caught the case where
a MIB (and thus the OID) was basically known, but not available (ie. because
there was no subagent connected serving it) and reported it with the wrong
exception name. Third, we didn't catch the case when an unknown OID was
specified (eg. an unknown MIB or a known MIB but an invalid OID within that
MIB).