Add first test cases for netsnmpagent module itself
These test cases so far yet deal with rather low level issues of the module
only, such as the expected behavior after instance creation, whether a
connection to a master snmpd agent inside a net-snmp test environment set
up with the netsnmptestenv module could be successfully established etc.

More important tests such as the behavior of the different types of SNMP
objects are yet to follow. For this reason, the included new TEST-MIB.txt
so far gets used only indirectly (to test when it becomes available to the
test environment's snmpd).