netsnmpagent: Make Table's value() method regard string lengths
When creating the dictionary returned by the Table class's value() method,
we so far assumed that the strings referenced through the "data" field in
netsnmp_table_data_set_storage structures would be zero byte-terminated, as
all C strings. However, this does not seem to be the case, as eg.
netsnmp_set_row_column() uses memcpy(), not strcpy(), ie. when setting a
string it is NOT zero byte-terminated. We thus need to regard "data_len"
when returning data to the Python world.

Strangely, this does not seem to be necessary with the simple scalar types
such as OctetString and DisplayString which are basically ctypes.c_char_p
objects as well...