Allow IpAddress objects to be used as table index. Add missing attributes to class IpAddress and adjust byte order as needed.
If IpAddress is used as table index, pass it to Net-SNMP
in host byte order. Else pass it in network byte order. See
https://sourceforge.net/p/net-snmp/bugs/2136 for details
why this is necessary.

Improve string representation of IpAddress in Table.value().

Note:
In IpAddress.cref() I don't store a reference to _cidx at "self".
This is only safe, if _cidx is consumed by a Net-SNMP function
that copies the value (like snmp_varlist_add_variable).
If it was used by Net-SNMP as raw pointer, Pythons reference
counting could delete the object while Net-SNMP still
wants to access it.

This changes have been tested on x86_64 (little endian)
and mips64 (big endian) architecture.