netsnmpagent: Rewrite SNMP variable type class instantiation
So far we used to have variable type-specific methods that return a
dictionary as input to a monstrous class factory implemented in the
@VarTypeClass decorator. Because it had to account for a wide amount of
variable types with all sorts of differences, it had a lot of branches
and still was not flexible enough for e.g. the IpAddress type.

This change moves all variable type specifics into a clean class
hierarchy housed in the new "netsnmpvartypes" module, bringing with it
advantages such as smaller code units, code inheritance that allows for
generic behavior implemented in root classes and specific behavior in
leaf classes etc.

Currently this class hierarchy looks like this:

                         _VarType
                           |  |
    _FixedSizeVarType <----'  '----> _MaxSizeVarType
      |                                |
      +-- Integer32                    `-- String
      +-- Unsigned32                       |
      +-- Counter32                        +-- OctetString
      +-- Counter64                        `-- DisplayString
      +-- TimeTicks
      `-- IpAddress

The classes have also been freed from any SNMP registration code. This
means that instantiating them directly is only useful when they're used
within a Table context.

For usage as scalar variables and, of course, to preserve compatibility,
we continue to provide methods inside the netsnmpAgent class that
instantiate VarType classes, set up the Net-SNMP watcher structure under
the hood and register them with the used agent instance. That is,

  integer = agent32.Integer("YOUR-MIB::yourInteger")

continues to work and behave as known.

This is implemented by having the netsnmpAgent class's __init__() parse
the "netsnmpvartypes" module for VarType classes and dynamically provide
methods for each that wrap class instantiation, Net-SNMP watcher setup
and registration with the agent instance's object registry.