netsnmpagent: Integrate IpAddress into @VarTypeClass class factory
This change resolves the separate IpAddress class in favor of the generic
@VarTypeClass class factory, extended to support IpAddress's
peculiarities.

As 7e28c23e showed, having a separate class is an invitation for bugs,
especially if it is copy-pasted as base for new variable types as in the
currently open pull requests.

The drawback on inflating the @VarTypeClass class factory with more and
more if-blocks will be addressed in a later commit.