Add support for Float variable type
The situation with Float (and even more so with Double) variables is
a bit unsatisfactory. To explain, let's hear the Net-SNMP mailing list
gods themselves.

In https://sourceforge.net/p/net-snmp/mailman/message/17349019/
Wes Hardaker on 2007-12-16 argued that:

  Generally, you should never use OPAQUEs. If you have an application
  specific binary blob of data, you should put it in an OCTET STRING.

  An OPAQUE is a special type that was actually deprecated in SMIv2, so
  it's not even supposed to be used.  The goal of it was to allow for
  extensibility but the SMI authors dropped it from suggested usage in
  second version of the language.

In https://sourceforge.net/p/net-snmp/mailman/message/18563538/
Dave Shield on 2008-02-11 argued that:

  If you want it to work, use ASN_OPAQUE_FLOAT.
  If you're not bothered, use ASN_FLOAT

  :-)

whereas in https://sourceforge.net/p/net-snmp/mailman/message/18563538/
Wes Hardaker on the same day added:

  ASN_FLOAT is not a legal SNMP data type.  ASN_OPAQUE_FLOAT encodes a
  float in a way that is legal for encoding in SNMP packets.  Note,
  however, that Net-SNMP is likely the only package that supports them.

So ASN_FLOAT is "not legal", but ASN_OPAQUE_FLOAT is Net-SNMP specific.
And in my experiments the latter actually worked, so we implement Float
support using ASN_OPAQUE_FLOAT, not ASN_FLOAT. You will have to include
the appropriate definition from NET-SNMP-TC in your MIB.

I did however not find any MIB that gives a definition for "Double",
not even in the Net-SNMP-TC. As far as I can see this seriously limits
its usability, so for now I'll not provide support for it in
python-netsnmpagent. Feel free to contact me if you can provide a
better solution.

Based on a merge request by Bill Randle <billr@neocat.org>.