netsnmptestenv: Improve datatype/data splitting
Floats (supported in an upcoming release) are a bit trickier to handle
with regard to parsing the snmpget output and splitting it into
"datatype" and "data" because they can embody double colons (:) as part
of the datatype (e.g. "Opaque: Float"). Just using rsplit alone would
break for TimeTicks which uses "0:00.00"-style data, so we use ": " to
rsplit the data. But then again, we have to take care of empty strings...